+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # adrez-mapper Revisions
 
+2020-11-25
+- add: contact.sortName where McAnser, Jay is translated to Jay McAnser
+
 2020-11-24
 - chg: empty telephone.type gets a default type
 - chg: empty email.type gets a default type